+++ src/civis/client.py
@@ -1,10 +1,18 @@
+from __future__ import annotations
+
 from functools import lru_cache
 import logging
+import textwrap
+from typing import TYPE_CHECKING
 
 import civis
 from civis.resources import generate_classes_maybe_cached
-from civis._utils import get_api_key
+from civis._utils import get_api_key, DEFAULT_RETRYING_STR
 from civis.response import _RETURN_TYPES, find, find_one
+
+if TYPE_CHECKING:
+    import collections
+    import tenacity
 
 
 _log = logging.getLogger(__name__)
@@ -38,15 +46,28 @@ class APIClient:
     force_refresh_api_spec : bool, optional
         Whether to force re-downloading the API spec,
         even if the cached version for the given API key hasn't expired.
+    retries : tenacity.Retrying, optional
+        Provide a :class:`tenacity.Retrying` instance for retries.
+        If ``None`` or not provided, the following default will be used:
+
+        .. code-block:: python
+
+           {default_retrying}
+
+        If you're providing a :class:`tenacity.Retrying` instance,
+        please note that you should leave the ``retry`` attribute unspecified,
+        because the conditions under which retries apply are pre-determined
+        -- see :ref:`retries` for details.
     """
 
     def __init__(
         self,
-        api_key=None,
-        return_type="snake",
-        api_version="1.0",
-        local_api_spec=None,
-        force_refresh_api_spec=False,
+        api_key: str | None = None,
+        return_type: str = "snake",
+        api_version: str = "1.0",
+        local_api_spec: collections.OrderedDict | str | None = None,
+        force_refresh_api_spec: bool = False,
+        retries: tenacity.Retrying | None = None,
     ):
         if return_type not in _RETURN_TYPES:
             raise ValueError(
@@ -54,7 +75,7 @@ class APIClient:
             )
         self._feature_flags = ()
         session_auth_key = get_api_key(api_key)
-        self._session_kwargs = {"api_key": session_auth_key}
+        self._session_kwargs = {"api_key": session_auth_key, "retrying": retries}
         self.last_response = None
 
         classes = generate_classes_maybe_cached(
@@ -66,6 +87,11 @@ class APIClient:
                 class_name,
                 cls(self._session_kwargs, client=self, return_type=return_type),
             )
+
+        # Don't create the `tenacity.Retrying` instance until we make the first
+        # API call with this `APIClient` instance.
+        # Once that happens, we keep re-using this `tenacity.Retrying` instance.
+        self._retrying = None
 
     @property
     def feature_flags(self):
@@ -322,3 +348,8 @@ class APIClient:
     def username(self):
         """The current user's username."""
         return self.users.list_me().username
+
+
+APIClient.__doc__ = APIClient.__doc__.format(
+    default_retrying=textwrap.indent(DEFAULT_RETRYING_STR.strip(), " " * 11).strip()
+)