+++ tests/test_utils.py
@@ -1,96 +1,21 @@
-from unittest import mock
-
-from requests import Request
-from requests import ConnectionError, ConnectTimeout
+import copy
 from datetime import datetime
 from math import floor
+from unittest import mock
 
-from civis._utils import maybe_get_random_name
-from civis._utils import retry
-from civis._utils import retry_request
+import tenacity
+from requests import Request
+from requests import ConnectionError
+
+from civis._utils import retry_request, DEFAULT_RETRYING
 from civis._utils import _RETRY_VERBS, _RETRY_CODES, _POST_RETRY_CODES
 
-import pytest
 
-
-@mock.patch("civis._utils.uuid")
-def test_maybe_random_name_random(mock_uuid):
-    random_name = "11111"
-    mock_uuid.uuid4.return_value = mock.Mock(hex=random_name)
-    assert maybe_get_random_name(None) == random_name
-
-
-def test_maybe_random_name_not_random():
-    given_name = "22222"
-    assert maybe_get_random_name(given_name) == given_name
-
-
-def test_io_no_retry():
-    @retry(ConnectionError, retries=4, delay=0.1)
-    def succeeds():
-        counter["i"] += 1
-        return "success"
-
-    counter = dict(i=0)
-    test_result = succeeds()
-
-    assert test_result == "success"
-    assert counter["i"] == 1
-
-
-def test_io_retry_once():
-    @retry(ConnectionError, retries=4, delay=0.1)
-    def fails_once():
-        counter["i"] += 1
-        if counter["i"] < 2:
-            raise ConnectionError("failed")
-        else:
-            return "success"
-
-    counter = dict(i=0)
-    test_result = fails_once()
-
-    assert test_result == "success"
-    assert counter["i"] == 2
-
-
-@mock.patch("civis.futures.time.sleep", side_effect=lambda x: None)
-def test_io_retry_limit_reached(m_sleep):
-    @retry(ConnectionError, retries=4, delay=0.1)
-    def always_fails():
-        counter["i"] += 1
-        raise ConnectionError("failed")
-
-    counter = dict(i=0)
-    pytest.raises(ConnectionError, always_fails)
-    assert counter["i"] == 5
-
-
-@mock.patch("civis.futures.time.sleep", side_effect=lambda x: None)
-def test_io_retry_multiple_exceptions(m_sleep):
-    @retry((ConnectionError, ConnectTimeout), retries=4, delay=0.1)
-    def raise_multiple_exceptions():
-        counter["i"] += 1
-        if counter["i"] == 1:
-            raise ConnectionError("one error")
-        elif counter["i"] == 2:
-            raise ConnectTimeout("another error")
-        else:
-            return "success"
-
-    counter = dict(i=0)
-    test_result = raise_multiple_exceptions()
-
-    assert test_result == "success"
-    assert counter["i"] == 3
-
-
-def test_io_retry_unexpected_exception():
-    @retry(ConnectionError, retries=4, delay=0.1)
-    def raise_unexpected_error():
-        raise ValueError("unexpected error")
-
-    pytest.raises(ValueError, raise_unexpected_error)
+def _get_retrying(retries: int):
+    retrying = copy.copy(DEFAULT_RETRYING)
+    stop = tenacity.stop_after_delay(600) | tenacity.stop_after_attempt(retries)
+    retrying.stop = stop
+    return retrying
 
 
 def test_no_retry_on_success():
@@ -112,7 +37,7 @@ def test_no_retry_on_success():
         )
         request = Request(**request_info)
         pre_request = session_context.prepare_request(request)
-        retry_request(verb, pre_request, session_context, 3)
+        retry_request(verb, pre_request, session_context, _get_retrying(3))
 
         assert session_context.send.call_count == expected_call_count
 
@@ -137,7 +62,7 @@ def test_no_retry_on_get_no_retry_failure():
         )
         request = Request(**request_info)
         pre_request = session_context.prepare_request(request)
-        retry_request(verb, pre_request, session_context, max_calls)
+        retry_request(verb, pre_request, session_context, _get_retrying(max_calls))
 
         assert session_context.send.call_count == expected_call_count
 
@@ -164,7 +89,7 @@ def test_retry_on_retry_eligible_failures(m_sleep):
 
             request = Request(**request_info)
             pre_request = session_context.prepare_request(request)
-            retry_request(verb, pre_request, session_context, max_calls)
+            retry_request(verb, pre_request, session_context, _get_retrying(max_calls))
 
             assert session_context.send.call_count == expected_call_count
 
@@ -191,7 +116,7 @@ def test_retry_on_retry_eligible_failures_lowercase_verbs(m_sleep):
 
             request = Request(**request_info)
             pre_request = session_context.prepare_request(request)
-            retry_request(verb, pre_request, session_context, max_calls)
+            retry_request(verb, pre_request, session_context, _get_retrying(max_calls))
 
             assert session_context.send.call_count == expected_call_count
 
@@ -214,7 +139,7 @@ def test_no_retry_on_post_success():
     )
     request = Request(**request_info)
     pre_request = session_context.prepare_request(request)
-    retry_request("post", pre_request, session_context, max_calls)
+    retry_request("post", pre_request, session_context, _get_retrying(max_calls))
 
     assert session_context.send.call_count == expected_call_count
 
@@ -240,7 +165,7 @@ def test_retry_on_retry_eligible_post_failures(m_sleep):
         )
         request = Request(**request_info)
         pre_request = session_context.prepare_request(request)
-        retry_request("post", pre_request, session_context, max_calls)
+        retry_request("post", pre_request, session_context, _get_retrying(max_calls))
 
         assert session_context.send.call_count == expected_call_count
 
@@ -266,7 +191,7 @@ def test_no_retry_on_connection_error():
 
         session_context.send.side_effect = ConnectionError()
         try:
-            retry_request(verb, pre_request, session_context, 3)
+            retry_request(verb, pre_request, session_context, _get_retrying(3))
         except ConnectionError:
             pass
 
@@ -314,7 +239,7 @@ def test_retry_respect_retry_after_headers():
         pre_request = session_context.prepare_request(request)
 
         start_time = datetime.now().timestamp()
-        retry_request(verb, pre_request, session_context, max_calls)
+        retry_request(verb, pre_request, session_context, _get_retrying(max_calls))
         end_time = datetime.now().timestamp()
         duration = end_time - start_time