+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## Unreleased
 
 ### Added
+- The new kwarg `retries` has been added to `civis.APIClient` so that
+  a `tenacity.Retrying` instance can be provided to customize retries. (#495)
+
 ### Changed
 ### Deprecated
 ### Removed